+++ Changelog.md
@@ -1,5 +1,5 @@
 # Change Log
-## [naive-version] - 2022-07-06
+## [naive-version] - 2022-07-08
 ## Added
 - initial structure
 - gitignore
@@ -13,3 +13,5 @@
 - skeletons postwallet and posttransaction as closure over rdb
 - Implemented genesis function
 - Implemented create wallet if not exists
+- Implemented post transactions, verify balance then execute and update balance, using optimistic log  - not thread safe
+- Implemented search wallets by tags